// src/app/models/arcctda-entity.ts
import { ArcctdaPKEntity } from './arcctda-pkentity';
export class ArcctdaEntity {
  arcctdaPKEntity: ArcctdaPKEntity;
  nombre: string;
  direccion: string;
  telefono: string;
  fax: string;
  direOfi: string;
  codiDepa: string;
  codiProv: string;
  codiDist: string;
  tipoDir: string;
  activo: string;
  tipoEnti: string;
  codSuc: string;
  noCliente1: string;
  estabSunat: string;
}